feat: Add MenuSelectTableBorderStyle component
The MenuSelectTableBorderStyle component has been added to the controls folder. This component is necessary to implement a specific table border style in the application.
