+++ package.json
@@ -88,6 +88,7 @@
     "@tiptap/extension-heading": "^2.0.0-beta.210",
     "@tiptap/extension-image": "^2.0.0-beta.210",
     "@tiptap/extension-table": "^2.0.0-beta.210",
+    "@tiptap/extension-table-cell": "^2.0.0-beta.210",
     "@tiptap/pm": "^2.0.0-beta.210",
     "@tiptap/react": "^2.0.0-beta.210",
     "react": "^16.8.0 || ^17.0.2 || ^18.0.0",
