+++ src/extensions/TableCellImproved.ts
@@ -1,26 +1,26 @@
 import { TableCell } from "@tiptap/extension-table-cell";
 
 const TableCellImproved = TableCell.extend({
-	addAttributes() {
-		return {
-			...this.parent?.(),
-			borderStyle: {
-				default: null,
-				renderHTML: (attributes) => {
-					if (!attributes.borderStyle) {
-						return {}
-					}
+  addAttributes() {
+    return {
+      ...this.parent?.(),
+      borderStyle: {
+        default: null,
+        renderHTML: (attributes) => {
+          if (!attributes.borderStyle) {
+            return {};
+          }
 
-					return {
-						style: `border-style: ${attributes.borderStyle}`,
-					}
-				},
-				parseHTML: (element) => {
-					return element.style.borderStyle.replace(/['"]+/g, '')
-				},
-			},
-		}
-	},
-})
+          return {
+            style: `border-style: ${attributes.borderStyle as string}`,
+          };
+        },
+        parseHTML: (element) => {
+          return element.style.borderStyle.replace(/['"]+/g, "");
+        },
+      },
+    };
+  },
+});
 
-export default TableCellImproved
+export default TableCellImproved;
