+++ src/icons/BorderStyleInset.tsx
@@ -1,9 +1,10 @@
 import { createSvgIcon } from "@mui/material";
 
 const BorderStyleInset = createSvgIcon(
-  <><path d="M5 5 H95" stroke="currentColor" />
-  <path d="M10 6 H90" stroke="currentColor" />
-  <path d="M15 7 H85" stroke="currentColor" />
+  <>
+    <path d="M5 5 H95" stroke="currentColor" />
+    <path d="M10 6 H90" stroke="currentColor" />
+    <path d="M15 7 H85" stroke="currentColor" />
   </>,
   "BorderStyleInset"
 );
