+++ src/icons/BorderStyleNone.tsx
@@ -1,7 +1,7 @@
 import { createSvgIcon } from "@mui/material";
 
 const BorderStyleNone = createSvgIcon(
-  <path d="M5 5 H95" stroke="none"/>,
+  <path d="M5 5 H95" stroke="none" />,
   "BorderStyleNone"
 );
