+++ src/icons/BorderStyleOutset.tsx
@@ -1,9 +1,10 @@
 import { createSvgIcon } from "@mui/material";
 
 const BorderStyleOutset = createSvgIcon(
-  <><path d="M5 5 H95" stroke="currentColor" />
-  <path d="M4 4 H96" stroke="currentColor" />
-  <path d="M6 6 H94" stroke="currentColor" />
+  <>
+    <path d="M5 5 H95" stroke="currentColor" />
+    <path d="M4 4 H96" stroke="currentColor" />
+    <path d="M6 6 H94" stroke="currentColor" />
   </>,
   "BorderStyleOutset"
 );
